+++ kor-common/src/main/java/com/sefford/kor/repositories/LruRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014 Saúl Díaz
+ * Copyright (C) 2017 Saúl Díaz
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@ package com.sefford.kor.repositories;
 import com.sefford.kor.repositories.interfaces.FastRepository;
 import com.sefford.kor.repositories.interfaces.RepoElement;
 import com.sefford.kor.repositories.interfaces.Repository;
-import com.sefford.kor.repositories.interfaces.Updateable;
 import com.sefford.kor.repositories.utils.LruCache;
 
 import java.util.ArrayList;
@@ -26,36 +25,43 @@ import java.util.Collection;
 import java.util.List;
 
 /**
- * Simple abstraction of a repository which provides a Memory interface using {@link android.support.v4.util.LruCache LRUCache}
- * instead of a {@link java.util.Map Map} backend.
- * <p/>
- * This allows the repository to be able to keep a steady footprint in memory instead of grabbing
- * all available space.
- * <p/>
- * Declaring additional memory repositories is as easy as extending this repository with particular
- * classes for Keys and Values.
+ * Repository which adds the capability to hold a limited number of items in cache.
+ * <p>
+ * The elements do use a {@link LruCache LruCache} as implementation and currently only
+ * supports a quantity-based LRU (not a size-based one)
  *
  * @author Saul Diaz <sefford@gmail.com>
  */
-public class LruMemoryRepository<K, V extends RepoElement<K> & Updateable<V>>
+public class LruRepository<K, V extends RepoElement<K>>
         implements Repository<K, V>, FastRepository<K, V> {
 
-    protected final LruCache<K, V> cache;
+    /**
+     * LRU which backs up the maximum amount of elements it can hold
+     */
+    protected final LruCache<K> lru;
+    /**
+     * Core {@link Repository Repository} which backs up the LRU repository itself
+     */
+    protected final Repository<K, V> repository;
 
-    public LruMemoryRepository(LruCache<K, V> cache) {
-        this.cache = cache;
+    /**
+     * Wraps a repository which can only hold maxSize elements
+     *
+     * @param repository Core repository to add the capability to
+     * @param maxSize    Max elements capable to add to the repository
+     */
+    public LruRepository(Repository<K, V> repository, int maxSize) {
+        this.repository = repository;
+        this.lru = new LruCache<>(maxSize);
     }
 
     @Override
     public V save(V element) {
-        V result = get(element.getId());
-        if (result == null) {
-            cache.put(element.getId(), element);
-            result = element;
-        } else {
-            result.update(element);
+        final K previous = lru.put(element.getId());
+        if (previous != null) {
+            repository.delete(previous, null);
         }
-        return result;
+        return repository.save(element);
     }
 
     @Override
@@ -68,12 +74,13 @@ public class LruMemoryRepository<K, V extends RepoElement<K> & Updateable<V>>
 
     @Override
     public boolean contains(K id) {
-        return cache.get(id) != null;
+        return lru.contains(id);
     }
 
     @Override
     public void delete(K id, V element) {
-        cache.remove(id);
+        lru.remove(id);
+        repository.delete(id, element);
     }
 
     @Override
@@ -85,12 +92,13 @@ public class LruMemoryRepository<K, V extends RepoElement<K> & Updateable<V>>
 
     @Override
     public V get(K id) {
-        return cache.get(id);
+        lru.refresh(id);
+        return repository.get(id);
     }
 
     @Override
     public Collection<V> getAll(Collection<K> ids) {
-        List<V> result = new ArrayList<V>();
+        final List<V> result = new ArrayList<>();
         for (final K id : ids) {
             final V element = get(id);
             if (element != null) {
@@ -102,17 +110,18 @@ public class LruMemoryRepository<K, V extends RepoElement<K> & Updateable<V>>
 
     @Override
     public void clear() {
-        cache.evictAll();
+        lru.clear();
+        repository.clear();
     }
 
     @Override
     public Collection<V> getAll() {
-        return cache.snapshot().values();
+        return repository.getAll();
     }
 
     @Override
     public boolean containsInMemory(K id) {
-        return contains(id);
+        return lru.contains(id);
     }
 
     @Override
@@ -137,6 +146,6 @@ public class LruMemoryRepository<K, V extends RepoElement<K> & Updateable<V>>
 
     @Override
     public boolean isAvailable() {
-        return cache != null;
+        return lru != null;
     }
 }