+++ kor-common/src/test/java/com/sefford/kor/repositories/BaseRepositoryTest.java
@@ -4,6 +4,7 @@ import com.sefford.kor.repositories.interfaces.RepoElement;
 import com.sefford.kor.repositories.interfaces.Repository;
 import com.sefford.kor.repositories.interfaces.Updateable;
 
+import com.sefford.kor.repositories.utils.TestElement;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -11,11 +12,14 @@ import org.mockito.Mock;
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyCollection;
 import static org.mockito.Matchers.anyLong;
@@ -31,9 +35,9 @@ import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 public class BaseRepositoryTest {
-    private static final long EXPECTED_FIRST_ID = 1;
-    private static final long EXPECTED_SECOND_ID = 2;
-    private static final long EXPECTED_THIRD_ID = 3;
+    private static final int EXPECTED_FIRST_ID = 1;
+    private static final int EXPECTED_SECOND_ID = 2;
+    private static final int EXPECTED_THIRD_ID = 3;
 
     @Mock
     private TestElement mockedElement1;
@@ -42,14 +46,14 @@ public class BaseRepositoryTest {
     @Mock
     private TestElement mockedElement3;
     @Mock
-    Repository<Long, TestElement> currentLevel;
+    Repository<Integer, TestElement> currentLevel;
     @Mock
-    Repository<Long, TestElement> nextLevel;
+    Repository<Integer, TestElement> nextLevel;
 
     private List<TestElement> elements = new ArrayList<TestElement>();
-    private List<Long> ids = new ArrayList<Long>();
+    private List<Integer> ids = new ArrayList<Integer>();
 
-    private BaseRepository<Long, TestElement> repository;
+    private BaseRepository<Integer, TestElement> repository;
 
     @Before
     public void setUp() throws Exception {
@@ -179,7 +183,7 @@ public class BaseRepositoryTest {
     public void testDeleteNoNextLevel() throws Exception {
         repository.delete(EXPECTED_FIRST_ID, mockedElement1);
         verify(currentLevel, times(1)).delete(EXPECTED_FIRST_ID, mockedElement1);
-        verify(nextLevel, never()).delete(anyLong(), (TestElement) any());
+        verify(nextLevel, never()).delete(anyInt(), (TestElement) any());
     }
 
     @Test
@@ -188,7 +192,7 @@ public class BaseRepositoryTest {
 
         repository.delete(EXPECTED_FIRST_ID, mockedElement1);
         verify(currentLevel, times(1)).delete(EXPECTED_FIRST_ID, mockedElement1);
-        verify(nextLevel, times(1)).delete(anyLong(), (TestElement) any());
+        verify(nextLevel, times(1)).delete(anyInt(), (TestElement) any());
     }
 
     @Test
@@ -288,11 +292,11 @@ public class BaseRepositoryTest {
 
     @Test
     public void testGetAllIdsNullElements() throws Exception {
-        when(currentLevel.contains(anyLong())).thenReturn(Boolean.TRUE);
+        when(currentLevel.contains(anyInt())).thenReturn(Boolean.TRUE);
 
         repository.getAll(ids);
 
-        verify(currentLevel, atLeastOnce()).get(anyLong());
+        verify(currentLevel, atLeastOnce()).get(anyInt());
     }
 
     @Test
@@ -315,12 +319,12 @@ public class BaseRepositoryTest {
 
     @Test
     public void testThreeTierResolvesGetCorrectly() throws Exception {
-        final Repository<Long, TestElement> firstLevel = mock(Repository.class);
-        final Repository<Long, TestElement> secondLevel = mock(Repository.class);
-        final Repository<Long, TestElement> thirdLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> firstLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> secondLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> thirdLevel = mock(Repository.class);
 
-        final BaseRepository<Long, TestElement> localLevel = new BaseRepositoryImpl(firstLevel, secondLevel);
-        final BaseRepository<Long, TestElement> repository = new BaseRepositoryImpl(localLevel, thirdLevel);
+        final BaseRepository<Integer, TestElement> localLevel = new BaseRepositoryImpl(firstLevel, secondLevel);
+        final BaseRepository<Integer, TestElement> repository = new BaseRepositoryImpl(localLevel, thirdLevel);
 
         when(firstLevel.get(EXPECTED_FIRST_ID)).thenReturn(null);
         when(firstLevel.isAvailable()).thenReturn(Boolean.TRUE);
@@ -333,12 +337,12 @@ public class BaseRepositoryTest {
 
     @Test
     public void testThreeTierResolvesContainsCorrectly() throws Exception {
-        final Repository<Long, TestElement> firstLevel = mock(Repository.class);
-        final Repository<Long, TestElement> secondLevel = mock(Repository.class);
-        final Repository<Long, TestElement> thirdLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> firstLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> secondLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> thirdLevel = mock(Repository.class);
 
-        final BaseRepository<Long, TestElement> localLevel = new BaseRepositoryImpl(firstLevel, secondLevel);
-        final BaseRepository<Long, TestElement> repository = new BaseRepositoryImpl(localLevel, thirdLevel);
+        final BaseRepository<Integer, TestElement> localLevel = new BaseRepositoryImpl(firstLevel, secondLevel);
+        final BaseRepository<Integer, TestElement> repository = new BaseRepositoryImpl(localLevel, thirdLevel);
 
         when(firstLevel.contains(EXPECTED_FIRST_ID)).thenReturn(Boolean.FALSE);
         when(firstLevel.isAvailable()).thenReturn(Boolean.TRUE);
@@ -350,20 +354,29 @@ public class BaseRepositoryTest {
         assertTrue(repository.contains(EXPECTED_FIRST_ID));
     }
 
-    class TestElement implements RepoElement<Long>, Updateable<TestElement> {
+    @Test
+    public void testTierOneDoesNotContainElementTierTwoFailsAndTierThreeResolves() {
+        final Repository<Integer, TestElement> firstLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> secondLevel = mock(Repository.class);
+        final Repository<Integer, TestElement> thirdLevel = mock(Repository.class);
 
-        @Override
-        public Long getId() {
-            return null;
-        }
+        final BaseRepository<Integer, TestElement> localLevel = new BaseRepositoryImpl(firstLevel, secondLevel);
+        final BaseRepository<Integer, TestElement> repository = new BaseRepositoryImpl(localLevel, thirdLevel);
 
-        @Override
-        public TestElement update(TestElement other) {
-            return this;
-        }
+        when(firstLevel.contains(EXPECTED_FIRST_ID)).thenReturn(Boolean.FALSE);
+        when(firstLevel.isAvailable()).thenReturn(Boolean.TRUE);
+        when(firstLevel.get(EXPECTED_FIRST_ID)).thenReturn(null);
+        when(secondLevel.contains(EXPECTED_FIRST_ID)).thenReturn(Boolean.TRUE);
+        when(secondLevel.isAvailable()).thenReturn(Boolean.FALSE);
+        when(secondLevel.get(EXPECTED_FIRST_ID)).thenReturn(null);
+        when(thirdLevel.isAvailable()).thenReturn(Boolean.TRUE);
+        when(thirdLevel.contains(EXPECTED_FIRST_ID)).thenReturn(Boolean.TRUE);
+        when(thirdLevel.get(EXPECTED_FIRST_ID)).thenReturn(mockedElement1);
+
+        assertThat(repository.get(EXPECTED_FIRST_ID), is(mockedElement1));
     }
 
-    class BaseRepositoryImpl extends BaseRepository<Long, TestElement> {
+    class BaseRepositoryImpl extends BaseRepository<Integer, TestElement> {
 
         /**
          * Creates a new instance of a BaseRepository with next level.
@@ -373,7 +386,7 @@ public class BaseRepositoryTest {
          * @param currentLevel Current Level of the Repository
          * @param nextLevel    Next Level of the Repository
          */
-        protected BaseRepositoryImpl(Repository<Long, TestElement> currentLevel, Repository<Long, TestElement> nextLevel) {
+        protected BaseRepositoryImpl(Repository<Integer, TestElement> currentLevel, Repository<Integer, TestElement> nextLevel) {
             super(currentLevel, nextLevel);
         }
     }